TST: add a simple test for the onetep calculator
This test provides a simple check that the onetep calculator is
running correctly. A single water molecule is generated and the
total energy is computed. The total energy is compared to a
known-good value. This value was computed using the abinit PAW
data sets for hydrogen and oxygen from: http://www.abinit.org/downloads/PAW2