ENH: parser for vasprun.xml files now reads all k-points as well as parameters
The previous version of the vasprun.xml parser read only the first k-point.
The modified version loops over all k-points and adds this information to
the SinglePointDFTCalculator object.

In addition the parser now collects the values of those computational
parameters that are defined in the vasp calculator and adds this information
to the `parameters` "dictionary" of the SinglePointDFTCalculator object.