MAINT: modified parsing of parameters in vasp_read_xml (!268)
In response to the discussion of merge request !268 I modified the
parsing of parameters in read_vasp_xml. While the previous version
imported the parameter list from the vasp calculator module, the
revised implementation simply parses all parameters available in
the vasprun.xml avoiding the cross import.

In addition, the "generation" block inside the "kpoints" block is
parsed and added to the "parameters" dictionary. This can be useful
in connection with database applications as it requires much less
storage space than the full k-points list.

Finally, the "parameters" dictionary is now an OrderedDict. This
ensures a consisting ordering.