MAINT: fixed a type issue that occured with some python versions
The line `search_range -= 1.0` in `find_optimal_cell_shape` was
only inserted to satisfy the pyflakes/pep8 test. `search_range`,
however, is in principle an integer numpy array, which caused
exceptions on some machines/python versions. I therefore merely
changed the line in question to `search_range = -1`.

At the same time I fix a minor error in the documentation as the
arguments of the `make_supercell` function at the end of
`defects.rst` were in the wrong order.