Full upgrade of phonons.py to make use of symmetries.
This is a port of the PHON code from D. Alfe. The Spacegroup.get_spacegroup method should be available (which uses spglib when available). 

The 'isfile' call is replaced with a parallel-sensitive one which does not block when e.g. running GPAW with MPI. A number of internal functions and added at the end, and a few other 'hidden' methods for internal use. 

Hope all goes OK (that's a big commit...).