Add toy model user01-user20 species from KIM API
(1) The function 'string2symbols', which parses strings into the
individual species present, needed to be adjusted because our
'userXX' species strings start with a lower-case letter and also
because the two-digit code following them does not indicate
multiplicity.

(2) In what I've done here, you *are* allowed to specify
multiplicity by placing one or more integers after the 'userXX' in
the string.  For instance, 'user013' is mapped to ['user01',
'user01', 'user01']

(3) Although the 'strings2symbols' function is part of the atoms
module in our fork, it looks like the upstream ASE repo has moved it
to the 'symbols' module.