CRYSTAL writer; provide convenient access to spglib refine_cell
It is quite easy to accidentally create redundant atom images when
creating the fort.34 file, as CRYSTAL permits this file to contain the
full set of atoms or just a minimal basis. As the check for images is
extremely tight, for most purposes it will be best to symmetrise the
structure to be safe. I have set the default to "False" to
minimise surprises when performing ASE optimisations, but there could
be good arguments for the default to lie either way.