ENH: add .gui extension to io.formats (CRYSTAL)
CRYSTAL17 fort.34 files are re-named with a .gui extension by the
runcrystal17 script. This script is provided by the CRYSTAL
maintainers, used in the tutorials and recommended for general use, so
it is quite reasonable that an ASE user will have some .gui files
lying around that they want to read in, even if ASE itself replaces
the runcryXX scripts. This extension is also acknowledged by the
external visualizer/editor DLV.