ENH: Use spglib to get minimal basis (asymmetric unit) for CRYSTAL
CRYSTAL will also accept an structure file with only a minimal set of
atoms defined; the rest are determined by symmetry operations. It
could be useful to work with these as this was the output format for
older versions of CRYSTAL - and the process might be used to write
structures to the INPUT/.d12 file instead of using an external
geometry file.

Not entirely convinced that this is working sensibly at the
moment. Between spglib and CRYSTAL there is a lot of converting
to/from primitive/conventional cells going on.