ENH: Read vasp KPOINTS files with line-mode content
While line-mode is not generally recommended for production
calculations, ASE users may occasionally want to load in data from
them. Here we provide some support to read KPOINTS files written in
this format. Rather than try to represent/store the line-mode
information, we immediately convert to a linearly-interpolated
path. This should be equivalent, and is already compatible with
the rest of the ASE calculator.

This addition prevents ASE from raising an error when 'restarting'
the Vasp calcuator from a Line-mode calculation directory. However,
the kpoints do not make their way into the _outputs_ information
obtained with calc.get_ibz_kpoints(). That information is read from
the vasprun.xml parser, which does not currently recognize the
corresponding block.