Remove kpts_params logic from vasprun.xml parser
This logic collects some metadata about how the k-point list was
assembled. However,

a) the information is not stored in the returned object, so is not
   recoverable by the user

b) the logic is broken when applied to a vasprun.xml from a
   "line-mode" KPOINTS file, as this uses a different format for the
   "generation" block.

Removing this logic allows VASP to restart from a "line-mode" run
without losing accesible functionality.