Support old Numpy by using scary functional python
Numpy 1.17 and above let you linspace between vectors, which is really
elegant and handy. But we still support older versions, so have to
flip the dimensions, interpolate over each axis individually and then
reassemble back into columns.

Including the logic for the elegant version so that one day when we're
no longer supporting Numpy<1.17 this can be spotted and cleaned up.