CRYSTAL symmetry writing: conventional cell only
For now we restrict the symmetry writing to conventional cell only by
checking the volume of the transformation matrix from spglib. Some
naive checks by displacing and rotating an FCC unit cell show that
CRYSTAL is at least not blowing up any more, so that's a start. More
stress testing needed before we can be confident in this. And still
need to implement the reverse process, which is the hard part...

It should be possible to also derive a set of working symmetry
operations, but it's not 100% clear how to separate the unit cell
change from the rototranslation to a standard
position/orientation. More work needed?