ENH: Add heat_flux property to calculators
What it says on the tin. This is mostly intended as a discussion
starting point (hello @flokno). For now, adding such a key means
that during MD, if calculators return this property, it at least
gets saved.

In general, we should probably discuss which units and conventions
to use, whether this is useful as a property, and so on.

An example implementation in, for instance, the LJ potential should
also be possible, as demonstration.