Use an example with a barrier instead of a dip
ASENEB method takes the tangent force of each image pair from the
image closest to the highest energy image. If the highest image is not
the middle image (as it won't be in a dip), then it will not place
symmetric tangent forces on the images at either side of the middle
image, and thus fail.

It should probably be mentioned somewhere that the combination of a
NEB path with a dip in energy and method='aseneb' and RNEB will fail.