wannier-fix-nnk: minor changes, still broken
Now the implementation for the Wannier centers is very similar to the
one in Wannier90, but it still does not work in the general case.
EX: works for Gamma-only calculation on single H atom, works on (3,3,3)
MP-grid calculation on the same system, does not work on (3,1,1)
MP-grid calculation on the same system.