small doc tweaks, deprecate Phonons.__call__
Some doc tweaks I didn't want to get mixed up with the class split.

Explicitly define Phonons.calculate as an abstractmethod.

"Symmetry" is ambiguous; made it clear that it means "matrix symmetry."

The usage of the term "lattice vectors" in the API is confusing to me
and I could not find similar usage elsewhere. (I would first assume this
term refers to cartesian vectors which correspond to translational
symmetries; or barring that, it could also mean the 3 cell vectors).
So I clarified that it is the integer coefficients of the cells.