split out PhononData - enough to write one test
Get enough of the refactor done to be able to write and execute one
test.

I had a really difficult time seeing how to keep support for Born charges
in here.  I'll do that in a future commit.  For now there are private
attributes on PhononsData where the Born charge data could be stored
(just so that the relevant parts of the computational code don't need to
be deleted yet), but they are always `None`.

Very likely, a lot of the code is now broken.  Many more tests are
needed.