ENH io.lammpsdata reads force field coefficients.
lammpsdata is able to store {Pair,Bond,Angle,Dihedral,Improper} Coeffs
sections from lammps data files.

These are stored in a dict in Atoms.info['coeffs'].

The number of {bond,angle,dihedral,improper} types read from the data
file are stored in Atom.info[types], only if the respective lines are
present in the datafile AND the atom_type supports it.