Add Displacements and VibrationsRunner
Vibrations is *brimming* with bugs and strange API choices and it is
next to impossible to backwards compatibly fix anything.

So VibrationsRunner is the new Vibrations, not held to the constraints
of backwards compatibility.

Displacements is me getting my foot in the door by abstracting these
computations over the set of displacements used.  This is a key step
towards eventually having a PhonopyDisplacements, and also helps
significantly clean things up by eliminating all of the handwritten
numerical derivative code that is currently duplicated across various
read() functions.