add raman calculator prototypes
They're in their own file now just to make them easier to find.

Methods that used to be on Displacement for working with files are now
on the corresponding Runner classes (which is my temporary(?) naming
scheme for what used to be the Raman Calculators).

At the moment they don't have methods to get RamanOutput.
It appears that StaticRamanCalculator was intended to produce files that
are in turn read by PlazcekStatic, but I'm not sure if we want to keep
things that way?