get rid of exkwargs and polkwargs
As alluded to in a recent commit message, the idea here is that
ExcitationListCalculators and StaticPolarizabilityCalculators can
hopefully be reusable.

I would imagine that the reason that exkwargs existed in the first
place is because the "traditional joined object for backwards
compatibility" comment on H2MorseExcitedStatesAndCalculator suggests
that, historically, the there was probably no concept of an ExcitationList
"calculator"; there was likely only ExcitationLists, and these
had compute() and read() methods that both had to be given kwargs.

But now that the calculator objects exist, those objects can already
contain all of the info that would've need to have been encoded in these
kwargs in the first place.