just do the unified cache thing already
Early on in this refactor, the runners were generally designed to each
have their own cache.  But then I realized that the cache is part of
the locking mechanism used to facilitate multiple running processes
in the original API, and this logic doesn't compose well with multiple
caches (basically, only the first cache's lock state matters, which
seemed weird and seemed to encourage having a single cache)

In response to realizing this I did.... something, and then now that
I'm looking at the code today I notice that nothing is actually doing
the lockfile thing at all anymore.

(I think I sort of partially moved towards a single cache but never
 fully fleshed it out)

So... where to handle this logic?

Well, for a while now I've had this idea of the different classes
registering themselves with some "single unified disp runner" that
merges the outputs from each class into one dictionary.  I've avoided
actually doing it until now because it sounds overengineered.
But, the thing is: I see a clear route forward if I introduce such a
thing, while the trail has been consistently muddy without it.

So there.  There it is.  DisplacementRunner is THE runner.
It's main responsibility is handling the unified cache and using it to
determine which computations are finished/locked/not started.

And now things like the Raman runners no longer have to return data that
includes 'forces' and whatever because this one runner will call all of
the calculate functions.

So can we like... finish this now?