Quick fix for two bugs in the Amber calculator:
1) When velocities are imported or exported from ASE <-> Amber,
rescaling was not applied for the unit conversion. Unfortunately
Amber does *not* respect the units listed in the netCDF file and
assumes its own conventions have been followed for restart files.
The ASE Amber calculator now thereofore adjusts to Amber's internal
units before writing.

2) It was previously very dangerous to run two instances of the
calculator in the same directory at the same time, because both
wrote to the mden and mdfrc files and tried to read them. It is
now possible to overwrite these values, if desired. Also improved
code for reading mden files to ensure the last value of the energy
is always read no matter what else has happened in the run.