Vibrations: move Frederiksen implementation from reader to data
The method='frederiksen' parameter is long-established in ASE
vibrations routines, and applies a sum-rule enforcing Newton's 3rd
law. There is no need for this to be applied at read-time however, as
it is fundamentally a post-processing correction to the Hessian which
was initially read in.

By treating it as post-processing we move the process to
VibrationsData; this makes it easier to examine the effect of this
correction by yielding a new VibrationsData object. The new method is
deliberately given quite a general name, so that other corrections can
be implemented (including combinations) in the same method.