Finite diff vibrations: ref_atoms can be guessed from displacements
In many practical cases it is

- fairly easy to determine the reference structure from the
  displacements, and

- mildly inconvenient to keep hold of the reference structure and pass
  it back into functions at analysis time.

This routine allows the reference atoms to be guessed by taking the
modal positions from the full set of displacments.

This may not be very memory efficient for large systems with
many displacements, as it constructs a 3-D array of all positions.
However, it should be quite robust against different inputs; as the
displacement set may include many different displacements along each
degree of freedom, we cannot guarantee a minimum number of
displacements that definitely have the information we need.
A more intelligent search is possible but might involve more slow
native-Python looping.