Working test of Frederiksen method
The key issue as identified by @glevi in !3012 is that the Vibrations
object applies translational symmetry after the force constants are
gathered. This means that even with the Frederiksen (i.e.
translational sum rule) correction applied during force-constant
gathering, the condition is likely to be broken in the results.

In order to test that the Frederiksen method did what was intended,
we can craft an input that will survive these subsequent
processes. It's a bit contrived but should do the job.